CRUD servicios, tiendas, bares, tipo de habitacion

+++ src/main/java/uniandes/edu/co/proyecto/controller/BarController.java
@@ -3,11 +3,8 @@ package uniandes.edu.co.proyecto.controller;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -23,7 +20,7 @@ public class BarController {
     private BarRepository barRepository;
 
     @GetMapping("/bares")
-    public String reservas(Model model){
+    public String bares(Model model){
         model.addAttribute("bares", barRepository.findAll());
         return "bar";
     }
@@ -41,26 +38,29 @@ public class BarController {
     }
 
     @GetMapping("/bares/{id}/edit")
-    public String barEditarForm(@PathVariable("id") String id, @ModelAttribute Bar bar){
+    public String barEditarForm(@PathVariable("id") String id, Model model){
         Optional<Bar> barExistente= barRepository.findById(id);
         if (barExistente.isPresent()) {
-            barRepository.save(bar);
-            return "barEditar";
+            Bar bar = barExistente.get();
+            model.addAttribute("bar", bar);
         }
-        else {
-            return "redirect:/bares";
-        }
+        return "BarEditar";
+        
     }
 
-    @DeleteMapping("/bares/{id}/delete")
-    public ResponseEntity<String> barEliminar(@PathVariable("id") String id){
-        try {
+    @PostMapping("/bares/{id}/edit/save")
+    public String editar(@PathVariable("id") String id, @ModelAttribute Bar bar){
+        barRepository.save(bar);
+        return "redirect:/bares";
+    }
+
+    @GetMapping("/bares/{id}/delete")
+    public String eliminarBar(@PathVariable String id){
+        Optional<Bar> barExistente = barRepository.findById(id);
+        if (barExistente.isPresent()) {
             barRepository.deleteById(id);
-            return new ResponseEntity<>("El bar ha sido eliminado", HttpStatus.OK);
         }
-        catch(Exception e) {
-            return new ResponseEntity<>("El bar no ha sido eliminado", HttpStatus.INTERNAL_SERVER_ERROR);
-        }
+        return "redirect:/bares";
     }
     
 }
